Correct the handling of shared/interleaved attribute arrays
We want to support both single use, interleaved and stacked (e.g. first N
positions, then N normals) arrays. Stacked seems to happen in glTF models
at least. We could split them up, but it seems more useful to support this.

However, that means we have to handle the offset differently, its a global
offset in the items in the shared array, not just in each group of items.