EffectComposer: Add need_source_texture option
Some passes are fine without swapping, but still need a the
source as a texture (i.e. can't be have the previous pass
rendered to the screen). For example a bloom pass will
use a temporary target so it can easily draw that to the
final screen, however it can't have the source be on the
screen as there is no texture for that.