EffectsComposer: Handle render passed on top of non-rendering passes right
If some previous pass was drawn to read_buffer then we can't draw a
non-clearing render pass (or other pass type with the new does_copy ==
FALSE) directly to the screen, because that would lose the previously
drawn content. Instead we draw to read_buffer and then the final
copy-to-screen will kick in if needed.