Drop use of GdkRGBA in favour of graphene_vec3_t
We never use the alpha anyway, so it is confusing to have it ther.

Also, it matches what we do for other things, for instance we
can use graphene to interpolate colors, etc.

Its also nice to drop dependencies on Gtk+ types in the API to
better handle possible differences between gtk3 and gtk4.