Fix assert if disabling all lights
If the previous frame had lights we set the various light
array to some size, and if the next frame didn't have *any* active
lights we were not resetting these uniform arrays due to the if in:

  if (priv->lights)
    material_apply_light_setup (m_uniforms, &priv->light_setup, FALSE);

This mean when we later updated the light info the arrays were the wrong
size (expected 0, but they were the old size).

The fix here is to just always set up the light uniforms when the material
needs it.

This fixes https://github.com/alexlarsson/gthree/issues/24