Stop using GdkGLContext APIs, make resources hang of GthreeRenderere
We were mostly using gdk_gl_context_get_current() to track when the
right gl context was active so we could immediately free stuff in
the resource unrealization case. We now instead make the
"current renderer" a thread local thing and set that where we're
guaranteed for it to be set (where we previously had asserts for this).

This also changes the resource APIs so that they are tied to the
renderer (which is really 1:1 to the gl context anyway).

With this change we no longer have any direct use of gdk/gtk outside
of GthreeArea, which will allow us to support multiple backends like
Gtk4, or even non-gtk use.