material: Avoid ownership issues in material per-renderer properties
We don't want the material properties to own the program because there
is no guarantee it will be freed, nor that when its freed the gl
context is right to unrealize it. So, we make this unowned, which is
fine because we only use it when the material is valid for that
renderer so will never happen if the renderer is alive (and thus the
program, because they live forever with the renderer).