PlaneHelper: Change polygon fill side calculation
For some reason even though we were using the same calculations as
three.js to calculate the "side" for the plane helper mesh object we
got different behaviour from three.js. I don't understand why, but this
change makes gthree and three.js render the same thing.

In particular, this is about the side of the semi-transparent "fill"
of the plane visualization of examples/clipping, versus the three.js
code in https://threejs.org/examples/#webgl_clipping_intersection